Add credentials.py Move get_mast_token function
