+++ jwql/edb/engineering_database.py
@@ -50,7 +50,7 @@ from bokeh.embed import components
 from bokeh.plotting import figure
 import numpy as np
 
-from jwql.utils.utils import get_mast_token
+from jwql.utils.credentials import get_mast_token
 from jwedb.edb_interface import query_single_mnemonic, query_mnemonic_info
